Промежуточный коммит. Для взятия данных с сервера
